fix: Corregir manejo de decimales y variantes en sistema de inventario
Cambios críticos:
- Cambiar parseInt a parseFloat para cantidades en APIs de inventario
- Corregir complete-reception para manejar variantes correctamente
- Agregar tolerancia (0.001) en comparación de discrepancias
- Agregar step="any" en inputs de cantidad para permitir decimales
- Incluir variant_id en stock movements para trazabilidad

Archivos modificados:
- unified-receive/route.ts: parseFloat para qtyReceived
- supplier-returns/route.ts: parseFloat para quantityBefore
- supplier-returns/lots/route.ts: parseFloat para quantityAvailable
- warehouses/route.ts: parseFloat para stock fields
- pos/orders/route.ts: parseFloat para FIFO lot processing
- complete-reception/route.ts: soporte completo de variantes
- ScannedProductCard.tsx: parseFloat y step="any" en inputs

+++ src/app/api/market/warehouses/route.ts
@@ -151,11 +151,11 @@ export async function GET(request: NextRequest) {
           isActive: row.is_active,
           allowNegativeStock: row.allow_negative_stock,
           productsCount: parseInt(row.products_count) || 0,
-          totalStock: parseInt(row.total_stock) || 0,
+          totalStock: parseFloat(row.total_stock) || 0,
           // Stock fields when productId is provided
-          stockOnHand: row.stock_on_hand != null ? parseInt(row.stock_on_hand) : 0,
-          stockReserved: row.stock_reserved != null ? parseInt(row.stock_reserved) : 0,
-          stockAvailable: row.stock_available != null ? parseInt(row.stock_available) : 0,
+          stockOnHand: row.stock_on_hand != null ? parseFloat(row.stock_on_hand) : 0,
+          stockReserved: row.stock_reserved != null ? parseFloat(row.stock_reserved) : 0,
+          stockAvailable: row.stock_available != null ? parseFloat(row.stock_available) : 0,
           createdAt: row.created_at,
           updatedAt: row.updated_at
         })),
